1) Support passing bootstraps and PSK via environment variables   - Uses "P2P_BOOTSTRAPS" for a space-separated list of multiaddresses   - Uses "P2P_PSK" for the PSK   - Resolves #13 2) Move StringsToMultiaddrs from p2pnode package to util package
